refactor: it removes code that was trasferred to _layout.tsx
