+++ mobile/app/_layout.tsx
@@ -1,0 +1,40 @@
+import { styled } from 'nativewind'
+import blurBg from '../src/assets/bg-blur.png'
+import Stripes from '../src/assets/stripes.svg'
+import { ImageBackground } from 'react-native'
+import {
+  useFonts,
+  Roboto_400Regular,
+  Roboto_700Bold,
+} from '@expo-google-fonts/roboto'
+import { BaiJamjuree_700Bold } from '@expo-google-fonts/bai-jamjuree'
+import { Slot, SplashScreen } from 'expo-router'
+import { StatusBar } from 'expo-status-bar'
+
+// it allows the stripes to use tailwindcss
+const StyledStripes = styled(Stripes)
+
+export default function Layout() {
+  const [hasLoadedFonts] = useFonts({
+    Roboto_400Regular,
+    Roboto_700Bold,
+    BaiJamjuree_700Bold,
+  })
+
+  if (!hasLoadedFonts) {
+    // loading image
+    return <SplashScreen />
+  }
+  return (
+    <ImageBackground
+      source={blurBg}
+      className="relative flex-1 bg-gray-900"
+      imageStyle={{ position: 'absolute', left: '-100%' }}
+    >
+      <StyledStripes className="absolute left-2" />
+      <StatusBar style="light" translucent />
+      {/* it shows page content */}
+      <Slot />
+    </ImageBackground>
+  )
+}
