+++ mobile/app/index.tsx
@@ -1,25 +1,11 @@
 import { StatusBar } from 'expo-status-bar'
-import { ImageBackground, Text, TouchableOpacity, View } from 'react-native'
-import {
-  useFonts,
-  Roboto_400Regular,
-  Roboto_700Bold,
-} from '@expo-google-fonts/roboto'
-
-import blurBg from '../src/assets/bg-blur.png'
-
-import Stripes from '../src/assets/stripes.svg'
+import { Text, TouchableOpacity, View } from 'react-native'
 import NLWLogo from '../src/assets/nlw-spacetime-logo.svg'
 import { makeRedirectUri, useAuthRequest } from 'expo-auth-session'
-import { BaiJamjuree_700Bold } from '@expo-google-fonts/bai-jamjuree'
-import { styled } from 'nativewind'
-import { useEffect } from 'react'
+import React, { useEffect } from 'react'
 import { api } from '../src/lib/api'
 import * as SecureStore from 'expo-secure-store'
 import { useRouter } from 'expo-router'
-
-// it allows the stripes to use tailwindcss
-const StyledStripes = styled(Stripes)
 
 const discovery = {
   authorizationEndpoint: 'https://github.com/login/oauth/authorize',
@@ -30,12 +16,6 @@ const discovery = {
 
 export default function App() {
   const router = useRouter()
-  const [hasLoadedFonts] = useFonts({
-    Roboto_400Regular,
-    Roboto_700Bold,
-    BaiJamjuree_700Bold,
-  })
-
   const [, response, signInWithGithub] = useAuthRequest(
     {
       clientId: 'a56c2ca0d794e5f14fe7',
@@ -69,17 +49,8 @@ export default function App() {
     }
   }, [response])
 
-  if (!hasLoadedFonts) {
-    return null
-  }
-
   return (
-    <ImageBackground
-      source={blurBg}
-      className="relative flex-1 items-center bg-gray-900 px-8 py-10"
-      imageStyle={{ position: 'absolute', left: '-100%' }}
-    >
-      <StyledStripes className="absolute left-2" />
+    <View className="flex-1 items-center px-8 py-10">
       <View className="flex-1 items-center justify-center gap-6">
         <NLWLogo />
         <View className="space-y-2">
@@ -105,6 +76,6 @@ export default function App() {
         Feito com 💜 no NLW da Rocketseat
       </Text>
       <StatusBar style="light" translucent />
-    </ImageBackground>
+    </View>
   )
 }
